clk/cdce925: Use PLL1 as Y1's input source
On Juno ProFPGA boards, the DPU secondary output's pxlclk is fed from
the Y1 output of a CDCEL949, with the divider's input muxed to PLL1.

A more complete patch would have both XTAL and PLL1 as clk parents
for Y1 and redo the logic so XTAL is preferred over PLL1 any time it
can provide the desired output, since changing PLL1 affects Y2, Y3,
and potentially Y5, depending on mux settings (with current code, Y5
is a non-issue). Such a patch would stand a chance of getting accepted
upstream (this one is guaranteed to be rejected), so we have to carry
less downstream.

Change-Id: I31365a7e797f6cdf82bf16b51d05ee5a9c9bbb98