PCI: mt7621: rework bringup sequence
Clean up the MT7621 PCIe root complex and associated ports bringup. Bring
down the ports that fail to initialise immediately, without remembering
slot numbers and other mental contorsions.

From the MT7621 fragmented Programming Guide and datasheet we can infer
that each PCIe port has an associated reset control bit in the RSTCTL
register and that the PERST mode bit in the GPIO_MODE register is for
controlling the reset of the whole root complex. Use that to drop support
for specifying individual reset GPIOs per port. There might be some
boards in the wild that have wired multiple GPIOs for resetting individual
ports, but I think they should also work with this version of the bringup.

Signed-off-by: Liviu Dudau <liviu@dudau.co.uk>