drm/komeda: Relax DT parsing for pipelines.
The YAML schema for Komeda only requires one pipeline to be present in
the device tree, but the code expects that all hardware capabilities
are described. It is perfectly valid to leave one pipeline disconnected
and disabled in the DT, so we need to relax the parsing code to allow
for that.

Cc: Brian Starket <brian.starkey@arm.com>
Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>