drm/arm/komeda: Avoid invalid pointer dereference in komeda_crtc_flush_and_wait_for_flip_done()
If we are waiting for a FLIP event without an input completion parameter
we create a temporary completion in komeda_crtc_flush_and_wait_for_flip_done()
and set kcrtc->disable_done pointer to it. If wait_for_completion_timeout()
finishes before the timeout we forget to clean up kcrtc->disable_done and
if there are any other spurious FLIP events happening, we end up dereferencing
an invalid pointer.

Properly declare the temp completion as _ONSTACK() and cleanup
kcrtc->disable_done regardless of the result of wait_for_completion_timeout().

Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>